+++ src/tracer/src/Aspect/JsonRpcAspect.php
@@ -74,7 +74,7 @@ class JsonRpcAspect extends AbstractAspect
             try {
                 $result = $proceedingJoinPoint->process();
             } catch (Throwable $e) {
-                if ($span = CT::get('tracer.span.' . static::class)) {
+                if (($span = CT::get('tracer.span.' . static::class)) && $this->switchManager->isEnable('exception') && ! $this->switchManager->isIgnoreException($e::class)) {
                     $span->setTag('error', true);
                     $span->log(['message', $e->getMessage(), 'code' => $e->getCode(), 'stacktrace' => $e->getTraceAsString()]);
                     CT::set('tracer.span.' . static::class, $span);