+++ docs/zh-tw/changelog.md
@@ -1,5 +1,37 @@
 # 版本更新記錄
 
+# v3.0.37 - 2023-09-22
+
+## 新增
+
+- [#6156](https://github.com/hyperf/hyperf/pull/6156) 為 `stringable` 元件增加了 `Str::replaceStart()` 等方法。
+
+## 最佳化
+
+- [#6154](https://github.com/hyperf/hyperf/pull/6154) 在使用驗證器元件時，如果原生方法 `json_validate` 存在，則使用其進行 `Json` 格式驗證。
+- [#6157](https://github.com/hyperf/hyperf/pull/6157) 在使用 `trace` 元件時，只記錄開啟記錄異常開關，並且不在忽略列表中的資訊。
+
+## 修復
+
+- [#6160](https://github.com/hyperf/hyperf/pull/6160) 修復當設定配置 `services.enable.register` 為 `false` 時，仍然會發布服務到服務中心的問題。
+- [#6162](https://github.com/hyperf/hyperf/pull/6162) 修復使用 `Crontab` 的時，當執行任務超過預設超時時間後，則不能很好的進行控制任務執行時機的問題。
+
+# v3.0.36 - 2023-09-15
+
+# 新增
+
+- [#6062](https://github.com/hyperf/hyperf/pull/6057) 為 `hyperf/tracer` 新增 `RequestTraceListener` 監聽器。
+- [#6143](https://github.com/hyperf/hyperf/pull/6143) 為 `hyperf/tracer` 增加 `ignore_exceptions` 配置。
+
+## 最佳化
+
+- [#6151](https://github.com/hyperf/hyperf/pull/6151) 最佳化 `hyperf/kafka` 元件中 `FailToConsume` 的觸發時機。
+
+## 修復
+
+- [#6117](https://github.com/hyperf/hyperf/pull/6117) 修復 `GRPC` 客戶端無法被複用的問題。
+- [#6146](https://github.com/hyperf/hyperf/pull/6146) 修復 `validateJson` 驗證規則在高於 PHP 8.0 版本後，無法正常使用的問題。
+
 # v3.0.35 - 2023-09-01
 
 ## 修復